# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JalSakshi API (v1 contracts only)",
    "description": "Contract-only surface generated for T05. Not a running backend - see docs/agent-workflow/handoff-T05.md.",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/sync/push": {
      "post": {
        "summary": "Idempotent sample event batch push",
        "operationId": "sync_push_v1_sync_push_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/commands": {
      "post": {
        "summary": "Role-checked, idempotent case command",
        "operationId": "case_command_v1_cases__case_id__commands_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/AssignCommand"
                  },
                  {
                    "$ref": "#/components/schemas/ReferToLabCommand"
                  },
                  {
                    "$ref": "#/components/schemas/RecordActionCommand"
                  },
                  {
                    "$ref": "#/components/schemas/AcceptActionCommand"
                  },
                  {
                    "$ref": "#/components/schemas/LinkRetestCommand"
                  },
                  {
                    "$ref": "#/components/schemas/RecordCommunicationCommand"
                  },
                  {
                    "$ref": "#/components/schemas/RequestClosureCommand"
                  },
                  {
                    "$ref": "#/components/schemas/CloseCommand"
                  },
                  {
                    "$ref": "#/components/schemas/ReopenCommand"
                  },
                  {
                    "$ref": "#/components/schemas/DismissCommand"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "assign": "#/components/schemas/AssignCommand",
                    "refer_to_lab": "#/components/schemas/ReferToLabCommand",
                    "record_action": "#/components/schemas/RecordActionCommand",
                    "accept_action": "#/components/schemas/AcceptActionCommand",
                    "link_retest": "#/components/schemas/LinkRetestCommand",
                    "record_communication": "#/components/schemas/RecordCommunicationCommand",
                    "request_closure": "#/components/schemas/RequestClosureCommand",
                    "close": "#/components/schemas/CloseCommand",
                    "reopen": "#/components/schemas/ReopenCommand",
                    "dismiss": "#/components/schemas/DismissCommand"
                  }
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AcceptActionCommand": {
        "properties": {
          "command_id": {
            "type": "string",
            "format": "uuid",
            "title": "Command Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "type": {
            "type": "string",
            "const": "accept_action",
            "title": "Type"
          },
          "payload": {
            "$ref": "#/components/schemas/AcceptActionPayload"
          }
        },
        "type": "object",
        "required": [
          "command_id",
          "expected_version",
          "type",
          "payload"
        ],
        "title": "AcceptActionCommand"
      },
      "AcceptActionPayload": {
        "properties": {
          "action_id": {
            "type": "string",
            "format": "uuid",
            "title": "Action Id"
          }
        },
        "type": "object",
        "required": [
          "action_id"
        ],
        "title": "AcceptActionPayload"
      },
      "AssignCommand": {
        "properties": {
          "command_id": {
            "type": "string",
            "format": "uuid",
            "title": "Command Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "type": {
            "type": "string",
            "const": "assign",
            "title": "Type"
          },
          "payload": {
            "$ref": "#/components/schemas/AssignPayload"
          }
        },
        "type": "object",
        "required": [
          "command_id",
          "expected_version",
          "type",
          "payload"
        ],
        "title": "AssignCommand"
      },
      "AssignPayload": {
        "properties": {
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          }
        },
        "type": "object",
        "required": [
          "owner_id"
        ],
        "title": "AssignPayload"
      },
...
      "CloseCommand": {
        "properties": {
          "command_id": {
            "type": "string",
            "format": "uuid",
            "title": "Command Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "type": {
            "type": "string",
            "const": "close",
            "title": "Type"
          },
          "payload": {
            "$ref": "#/components/schemas/ClosePayload"
          }
        },
        "type": "object",
        "required": [
          "command_id",
          "expected_version",
          "type",
          "payload"
        ],
        "title": "CloseCommand"
      },
      "ClosePayload": {
        "properties": {
          "verified_report_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Report Id"
          },
          "verified_report_exemption_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Report Exemption Reason"
          },
          "retest_sample_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retest Sample Id"
          },
          "retest_exemption_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Retest Exemption Reason"
          },
          "action_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Action Ids"
          },
          "action_exemption_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Exemption Reason"
          },
          "communication_id": {
            "type": "string",
            "format": "uuid",
            "title": "Communication Id"
          },
          "disposition": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Disposition"
          },
          "policy_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Policy Version"
          }
        },
        "type": "object",
        "required": [
          "communication_id",
          "disposition",
          "policy_version"
        ],
        "title": "ClosePayload",
        "description": "api-contracts.md: 'close includes verified_report_id,\nretest_sample_id or authorized exemption, action IDs or exemption,\ncommunication_id, disposition and policy_version.' The or-exemption\npattern is enforced below, not just documented - this is what makes\nacceptance criterion 1 ('command discriminators reject invalid data')\nreal rather than type-checking alone."
      },
...
      "DismissCommand": {
        "properties": {
          "command_id": {
            "type": "string",
            "format": "uuid",
            "title": "Command Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "type": {
            "type": "string",
            "const": "dismiss",
            "title": "Type"
          },
          "payload": {
            "$ref": "#/components/schemas/DismissPayload"
          }
        },
        "type": "object",
        "required": [
          "command_id",
          "expected_version",
          "type",
          "payload"
        ],
        "title": "DismissCommand"
      },
...
      "LinkRetestCommand": {
        "properties": {
          "command_id": {
            "type": "string",
            "format": "uuid",
            "title": "Command Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "type": {
            "type": "string",
            "const": "link_retest",
            "title": "Type"
          },
          "payload": {
            "$ref": "#/components/schemas/LinkRetestPayload"
          }
        },
        "type": "object",
        "required": [
          "command_id",
          "expected_version",
          "type",
          "payload"
        ],
        "title": "LinkRetestCommand"
      },
      "LinkRetestPayload": {
        "properties": {
          "retest_sample_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retest Sample Id"
          }
        },
        "type": "object",
        "title": "LinkRetestPayload"
      },
...
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "field_errors": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Errors"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "request_id",
          "retryable"
        ],
        "title": "ProblemDetail"
      },
...
      "PushRequest": {
        "properties": {
          "device_id": {
            "type": "string",
            "format": "uuid",
            "title": "Device Id"
          },
          "events": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/SampleCreateEvent"
                },
                {
                  "$ref": "#/components/schemas/SampleCorrectEvent"
                }
              ],
              "discriminator": {
                "propertyName": "kind",
                "mapping": {
                  "sample.correct": "#/components/schemas/SampleCorrectEvent",
                  "sample.create": "#/components/schemas/SampleCreateEvent"
                }
              }
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "events"
        ],
        "title": "PushRequest",
        "description": "POST /v1/sync/push envelope. Batch bound (50) matches\napi-contracts.md 'Common rules' - not an arbitrary number picked here."
      },
      "RecordActionCommand": {
        "properties": {
          "command_id": {
            "type": "string",
            "format": "uuid",
            "title": "Command Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "type": {
            "type": "string",
            "const": "record_action",
            "title": "Type"
          },
          "payload": {
            "$ref": "#/components/schemas/RecordActionPayload"
          }
        },
        "type": "object",
        "required": [
          "command_id",
          "expected_version",
          "type",
          "payload"
        ],
        "title": "RecordActionCommand"
      },
      "RecordActionPayload": {
        "properties": {
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Description"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At"
          }
        },
        "type": "object",
        "required": [
          "description",
          "owner_id",
          "due_at"
        ],
        "title": "RecordActionPayload"
      },
      "RecordCommunicationCommand": {
        "properties": {
          "command_id": {
            "type": "string",
            "format": "uuid",
            "title": "Command Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "type": {
            "type": "string",
            "const": "record_communication",
            "title": "Type"
          },
          "payload": {
            "$ref": "#/components/schemas/RecordCommunicationPayload"
          }
        },
        "type": "object",
        "required": [
          "command_id",
          "expected_version",
          "type",
          "payload"
        ],
        "title": "RecordCommunicationCommand"
      },
      "RecordCommunicationPayload": {
        "properties": {
          "channel": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Channel"
          },
          "template_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Template Version"
          },
          "audience_description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Audience Description"
          }
        },
        "type": "object",
        "required": [
          "channel",
          "template_version",
          "audience_description"
        ],
        "title": "RecordCommunicationPayload"
      },
      "ReferToLabCommand": {
        "properties": {
          "command_id": {
            "type": "string",
            "format": "uuid",
            "title": "Command Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "type": {
            "type": "string",
            "const": "refer_to_lab",
            "title": "Type"
          },
          "payload": {
            "$ref": "#/components/schemas/ReferToLabPayload"
          }
        },
        "type": "object",
        "required": [
          "command_id",
          "expected_version",
          "type",
          "payload"
        ],
        "title": "ReferToLabCommand"
      },
      "ReferToLabPayload": {
        "properties": {
          "lab_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Lab Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "ReferToLabPayload"
      },
      "ReopenCommand": {
        "properties": {
          "command_id": {
            "type": "string",
            "format": "uuid",
            "title": "Command Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "type": {
            "type": "string",
            "const": "reopen",
            "title": "Type"
          },
          "payload": {
            "$ref": "#/components/schemas/ReopenPayload"
          }
        },
        "type": "object",
        "required": [
          "command_id",
          "expected_version",
          "type",
          "payload"
        ],
        "title": "ReopenCommand"
      },
      "ReopenPayload": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "ReopenPayload"
      },
      "RequestClosureCommand": {
        "properties": {
          "command_id": {
            "type": "string",
            "format": "uuid",
            "title": "Command Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "type": {
            "type": "string",
            "const": "request_closure",
            "title": "Type"
          },
          "payload": {
            "$ref": "#/components/schemas/RequestClosurePayload"
          }
        },
        "type": "object",
        "required": [
          "command_id",
          "expected_version",
          "type",
          "payload"
        ],
        "title": "RequestClosureCommand"
      },
      "RequestClosurePayload": {
        "properties": {
          "disposition": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Disposition"
          }
        },
        "type": "object",
        "title": "RequestClosurePayload"
      },
      "SampleCorrectEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "kind": {
            "type": "string",
            "const": "sample.correct",
            "title": "Kind"
          },
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "payload": {
            "$ref": "#/components/schemas/SampleCorrectionSampleV1"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "kind",
          "payload"
        ],
        "title": "SampleCorrectEvent"
      },
...
      "SampleCreateEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "kind": {
            "type": "string",
            "const": "sample.create",
            "title": "Kind"
          },
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "payload": {
            "$ref": "#/components/schemas/CanonicalSampleV1"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "kind",
          "payload"
        ],
        "title": "SampleCreateEvent"
      },
...
    }
  }
}